Rename plugin sources
For improved consistency; also move modified Yosys code to `/third_party`.
